acv/pr: ensure vod database is cleaned inbetween drs tests

+++ drs/src/core/com/db4o/drs/versant/VodReplicationProvider.java
@@ -92,6 +92,9 @@ public class VodReplicationProvider implements TestableReplicationProviderInside
 	}
 
 	public void deleteAllInstances(Class clazz) {
+		if(!_jdo.isKnownClass(clazz)) {
+			return;
+		}
 		_jdo.deleteAll(clazz);
 	}
